Added files via upload
Battlefield.java is a class to create the battlefield, with two players. This class also allows the players to move back & forth and checks for collisions. I left main in the java file because that how I was testing it. I'm pretty sure I've met all the requirements, if I didn't let me know :D